# docs/api/openapi-v1.json
{
  "components": {
    "schemas": {
      "AreaData": {
        "properties": {
          "data": {
            "additionalProperties": true,
            "title": "Data",
            "type": "object"
          }
        },
        "required": [
          "data"
        ],
        "title": "AreaData",
        "type": "object"
      },
...
      "ConfigFile": {
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "file_type": {
            "default": "markdown",
            "title": "File Type",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "path",
          "content"
        ],
        "title": "ConfigFile",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "McpServerUpdate": {
        "properties": {
          "servers": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Servers",
            "type": "array"
          }
        },
        "required": [
          "servers"
        ],
        "title": "McpServerUpdate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "StudioCookie": {
        "in": "cookie",
        "name": "boerdi_studio_auth",
        "type": "apiKey"
      },
      "StudioKey": {
        "in": "header",
        "name": "X-Studio-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "boerdi-chat backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/config/data/{area}": {
      "get": {
        "description": "The area's stored document as JSON — the counterpart /schema needs.\n\n``/config/file`` returns YAML *text*, which a form cannot bind to. An area\nthat has a model but no row yields ``{}`` rather than 404: the schema is\nthere, so the form renders defaults and the first save creates it.\n\n``type`` says which file the raw editor must ask for, served from the same\n``seed_io.is_md_area`` predicate ``read_config_file`` uses to decide how to\nserialize. The studio therefore never guesses it from the document shape,\nand the two sides cannot disagree about a document.",
        "operationId": "get_area_data_api_config_data__area__get",
        "parameters": [
          {
            "in": "path",
            "name": "area",
            "required": true,
            "schema": {
              "title": "Area",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Area Data Api Config Data  Area  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "StudioKey": []
          }
        ],
        "summary": "Get Area Data",
        "tags": [
          "config"
        ]
      },
      "put": {
        "description": "Replace the area's document (9-3a). ``data`` is the WHOLE document.\n\nGET/PUT are a read-modify-write pair, and that is deliberate. The area\nmodels pin only part of the tree — measured against the ALT config, 357\ndata paths sit below or beside a pinned key without being pinned\nthemselves (``01-base/policy`` -> ``rules[*].effect.disclaimer``,\n``01-base/classify-overrides`` -> ``pattern_disambiguators_legacy[*]``).\nA schema-driven form therefore edits a *copy of the whole document* and\nsubmits it whole; the unpinned parts ride along untouched.\n\nMerging server-side was the other candidate and is worse: a deep merge\ncannot express deletion (dropping a list entry or a whole section would be\nimpossible), and a shallow one would only have protected the top level —\nwhich is not where the unpinned keys are. Replace also matches every typed\narea endpoint, so the store has one write semantic, not two.\n\nValidation is a gate, not a transform: the submitted *raw* dict is\npersisted, never the model dump — dumping would inject defaults for every\nabsent optional field and rewrite the document the editor saw.",
        "operationId": "update_area_data_api_config_data__area__put",
        "parameters": [
          {
            "in": "path",
            "name": "area",
            "required": true,
            "schema": {
              "title": "Area",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AreaData"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Update Area Data Api Config Data  Area  Put",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "StudioKey": []
          }
        ],
        "summary": "Update Area Data",
        "tags": [
          "config"
        ]
      }
    },
...
    "/api/config/file": {
      "delete": {
        "operationId": "delete_config_file_api_config_file_delete",
        "parameters": [
          {
            "in": "query",
            "name": "path",
            "required": true,
            "schema": {
              "title": "Path",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Delete Config File Api Config File Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "StudioKey": []
          }
        ],
        "summary": "Delete Config File",
        "tags": [
          "config"
        ]
      },
      "get": {
        "operationId": "get_config_file_api_config_file_get",
        "parameters": [
          {
            "in": "query",
            "name": "path",
            "required": true,
            "schema": {
              "title": "Path",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Config File Api Config File Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "StudioKey": []
          }
        ],
        "summary": "Get Config File",
        "tags": [
          "config"
        ]
      },
      "put": {
        "description": "Persist raw file text.\n\nMalformed YAML is the caller's mistake, not a server fault: ``safe_load``\nraises ``yaml.YAMLError`` (a subclass of neither ``ValueError`` nor\nanything else caught here), so before 9-3 a typo in the studio's raw editor\nproduced a 500 with a stack trace instead of a message the editor could\nact on. The parser's own message names the line and column.",
        "operationId": "update_config_file_api_config_file_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigFile"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Update Config File Api Config File Put",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "StudioKey": []
          }
        ],
        "summary": "Update Config File",
        "tags": [
          "config"
        ]
      }
    },
    "/api/config/files": {
      "get": {
        "operationId": "list_config_files_api_config_files_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response List Config Files Api Config Files Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "StudioKey": []
          }
        ],
        "summary": "List Config Files",
        "tags": [
          "config"
        ]
      }
    },
...
    "/api/config/mcp-servers": {
      "get": {
        "operationId": "get_mcp_servers_api_config_mcp_servers_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Get Mcp Servers Api Config Mcp Servers Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "StudioKey": []
          }
        ],
        "summary": "Get Mcp Servers",
        "tags": [
          "config"
        ]
      },
      "put": {
        "description": "Replace the MCP server registry.\n\nEvery non-primary URL is SSRF-checked before anything is persisted (ALT audit\nT8, stored SSRF): the backend POSTs to each enabled server on every chat turn,\nso an internal URL accepted here would fire once per conversation.\n\nThe primary (``wlo-mcp``) is exempt: its URL comes from ``MCP_SERVER_URL`` and\nmay deliberately point inside the network, and ``save_mcp_servers`` strips it\nbefore writing regardless.",
        "operationId": "update_mcp_servers_api_config_mcp_servers_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/McpServerUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Update Mcp Servers Api Config Mcp Servers Put",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "StudioKey": []
          }
        ],
        "summary": "Update Mcp Servers",
        "tags": [
          "config"
        ]
      }
    },
    "/api/config/mcp-servers/discover": {
      "post": {
        "description": "Handshake an MCP server once and list its tools, without registering it.\n\nThe SSRF guard sits in *front* of the egress, not behind a stored value like\nthe PUT above: this endpoint dials whatever URL it is handed, so an internal\nURL must be rejected before the handshake fires rather than before a later\nwrite (ALT audit T8/T9). A transport failure is a 502 — the server is\nreachable-or-not from here, it is not the caller's bad request.",
        "operationId": "discover_mcp_tools_api_config_mcp_servers_discover_post",
        "parameters": [
          {
            "in": "query",
            "name": "url",
            "required": false,
            "schema": {
              "default": "",
              "title": "Url",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Discover Mcp Tools Api Config Mcp Servers Discover Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "StudioKey": []
          }
        ],
        "summary": "Discover Mcp Tools",
        "tags": [
          "config"
        ]
      }
    },
...
    "/api/config/schema/{area}": {
      "get": {
        "description": "JSON schema of the area's pydantic model (P2-6) — grouped file keys\n(03-patterns/*, 04-personas/*) resolve to their group model.",
        "operationId": "get_area_schema_api_config_schema__area__get",
        "parameters": [
          {
            "in": "path",
            "name": "area",
            "required": true,
            "schema": {
              "title": "Area",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Area Schema Api Config Schema  Area  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "StudioKey": []
          }
        ],
        "summary": "Get Area Schema",
        "tags": [
          "config"
        ]
      }
    },
...
    "/api/eval/config": {
      "get": {
        "description": "Current personas + intents from the active chatbot config.",
        "operationId": "eval_config_api_eval_config_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Eval Config Api Eval Config Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "StudioKey": []
          }
        ],
        "summary": "Eval Config",
        "tags": [
          "eval"
        ]
      }
    },
...
  }
}